backend/microservices/ai-brain/requirements.txt
# AI Brain Service - Python Dependencies

# Flask Framework
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0

# Machine Learning & NLP
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3

# OpenAI Integration
openai==1.3.0

# Database
pymongo==4.6.0
redis==5.0.1

# Utilities
python-dateutil==2.8.2
requests==2.31.0

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
